% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport-gsDesign.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{gsDesign}
\alias{sfBetaDist}
\alias{sfCauchy}
\alias{sfExponential}
\alias{sfExtremeValue}
\alias{sfExtremeValue2}
\alias{sfGapped}
\alias{sfHSD}
\alias{sfLDOF}
\alias{sfLDPocock}
\alias{sfLinear}
\alias{sfLogistic}
\alias{sfNormal}
\alias{sfPoints}
\alias{sfPower}
\alias{sfStep}
\alias{sfTDist}
\alias{sfTrimmed}
\alias{sfTruncated}
\alias{sfXG1}
\alias{sfXG2}
\alias{sfXG3}
\alias{gsBoundSummary}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{gsDesign}{\code{\link[gsDesign:gsBoundSummary]{gsBoundSummary()}}, \code{\link[gsDesign:gsDesign]{gsDesign()}}, \code{\link[gsDesign:sfBetaDist]{sfBetaDist()}}, \code{\link[gsDesign:sfCauchy]{sfCauchy()}}, \code{\link[gsDesign:sfExponential]{sfExponential()}}, \code{\link[gsDesign:sfExtremeValue]{sfExtremeValue()}}, \code{\link[gsDesign:sfExtremeValue2]{sfExtremeValue2()}}, \code{\link[gsDesign:sfGapped]{sfGapped()}}, \code{\link[gsDesign:sfHSD]{sfHSD()}}, \code{\link[gsDesign:sfLDOF]{sfLDOF()}}, \code{\link[gsDesign:sfLDPocock]{sfLDPocock()}}, \code{\link[gsDesign:sfLinear]{sfLinear()}}, \code{\link[gsDesign:sfLogistic]{sfLogistic()}}, \code{\link[gsDesign:sfNormal]{sfNormal()}}, \code{\link[gsDesign:sfPoints]{sfPoints()}}, \code{\link[gsDesign:sfPower]{sfPower()}}, \code{\link[gsDesign:sfStep]{sfStep()}}, \code{\link[gsDesign:sfTDist]{sfTDist()}}, \code{\link[gsDesign:sfTrimmed]{sfTrimmed()}}, \code{\link[gsDesign:sfTruncated]{sfTruncated()}}, \code{\link[gsDesign:sfXG1]{sfXG1()}}, \code{\link[gsDesign:sfXG2]{sfXG2()}}, \code{\link[gsDesign:sfXG3]{sfXG3()}}}
}}

