% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsDesign-reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{sfBetaDist}
\alias{sfCauchy}
\alias{sfExponential}
\alias{sfExtremeValue}
\alias{sfExtremeValue2}
\alias{sfGapped}
\alias{sfHSD}
\alias{sfLDOF}
\alias{sfLDPocock}
\alias{sfLinear}
\alias{sfLogistic}
\alias{sfNormal}
\alias{sfPoints}
\alias{sfPower}
\alias{sfStep}
\alias{sfTDist}
\alias{sfTrimmed}
\alias{sfTruncated}
\alias{sfXG1}
\alias{sfXG2}
\alias{sfXG3}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{gsDesign}{\code{\link[gsDesign:sfDistribution]{sfBetaDist}}, \code{\link[gsDesign:sfDistribution]{sfCauchy}}, \code{\link[gsDesign]{sfExponential}}, \code{\link[gsDesign:sfDistribution]{sfExtremeValue}}, \code{\link[gsDesign:sfDistribution]{sfExtremeValue2}}, \code{\link[gsDesign:sfSpecial]{sfGapped}}, \code{\link[gsDesign]{sfHSD}}, \code{\link[gsDesign]{sfLDOF}}, \code{\link[gsDesign:sfLDOF]{sfLDPocock}}, \code{\link[gsDesign]{sfLinear}}, \code{\link[gsDesign:sfDistribution]{sfLogistic}}, \code{\link[gsDesign:sfDistribution]{sfNormal}}, \code{\link[gsDesign]{sfPoints}}, \code{\link[gsDesign]{sfPower}}, \code{\link[gsDesign:sfLinear]{sfStep}}, \code{\link[gsDesign]{sfTDist}}, \code{\link[gsDesign:sfSpecial]{sfTrimmed}}, \code{\link[gsDesign:sfSpecial]{sfTruncated}}, \code{\link[gsDesign:sfXG]{sfXG1}}, \code{\link[gsDesign:sfXG]{sfXG2}}, \code{\link[gsDesign:sfXG]{sfXG3}}}
}}

